Remove stray fake time side effect in data provider
This caused Wikibase CI to be permanently red. This is particularly
annoying as this is a side effect in a data provider which are executed
even when the test suite is merely loaded but without the tests actually
being run. That way they got Wikibase tests to fail.

The offending line was introduced in
Iccc0c39e013679f15f2e62738c56df09ba56c494

Bug: T259785
Change-Id: I0df26d6af56ed18eac836929fd469b4a637d4902