Guard against 'Widget not found' error
Unlike other similar pages such as Special:MovePage, the reason field
and checkboxes are not always present here. Notably they will not exist
when navigating to Special:Undelete/existingpage or when the user has
permission to see/browse deleted revisions but no permission to undelete.

When the widget does not exist because of these server-side checks,
attempting the JS infusion will lead to a 'Widget not found' error.

Bug: T259916
Change-Id: I54be770f229cdacfae81d10fccd3cf2931df3a27