ResourceLoaderOOUIImageModule: loadOOUIDefinition() may return false
loadOOUIDefinition() may return false when an extension is (ab)using
the ResourceLoaderOOUIImageModule class to define custom icons (e.g.
VisualEditor), and it doesn't include definitions for all of the
available themes, perhaps because your skin is using a custom theme.

As noted in a code comment elsewhere in this file: "that's okay, it
will just use the defaults".

Bug: T260059
Change-Id: I5468906b5efc4b727c6967ab5958fa6657e355ee