SpecialBlock: Show error if a block could not be inserted or found
If a block cannot be inserted, it is assumed that there is an
existing block against the target. That block is then retrieved
from a replica database using DatabaseBlock::newFromTarget. We are
seeing errors as a result of assuming that newFromTarget always
returns a block object in this situation.

Sometimes a block could not be inserted because there is an existing
block, but the existing block could not be retrieved. This could
happen if:
* there is an existing block but it is not in the replicas yet
* an existing block was removed after the insert attempt, but before
  the retrieve attempt

Check whether DatabaseBlock::newFromTarget returns a block object,
and display a message explaining the situation if not.

Bug: T259212
Change-Id: I1737e3a69748ebaa743e87b185ba1e3b92afec8c