Sanitizer: Truncate IDs to a reasonable length; deprecate escapeIdReferenceList
Overly-long anchors can cause OOMs later on during TOC processing, and
are needless.

The method Sanitizer::escapeIdReferenceList() is also deprecated in
this patch, since it is a way to get around the ID length limit and
appears to be unused outside the Sanitizer class.  Since the use
within Sanitizer (for ARIA attributes) appears safe, we'll just make
this private in a future release and avoid the potential that someone
will misuse this.

Bug: T251506
Change-Id: Ifce057b0c436eabec310f812394e86ee7123e7c8