Move notification overlay outside $content
Move the mw.notify overlay to be appended to the document
body rather than inside the .mw-body element. This makes
it overlay everything, and so avoid being chopped off
when skins set position:relative on .mw-body.

Because this moves the overlay, it has to also vertically
re-position its contents (the notification area)
to match .mw-body.

Bug: T260413
Bug: T260262
Bug: T260432
Change-Id: Iacea9f65ed4be5afae377cfbf423b1c700f4367a