Allow REST API POST handlers to opt out of mandatory sqlite locking
This is a follow up to T93097, which worked around a limitation of
SQLite 3.8+ which prevents it from upgrading a read transaction to a write
transaction.  Our heuristic is that HTTP POST requests are going to
eventually become DB writes, and so we take the SQLite lock early.

However, Parsoid POST requests don't have any side effects on the DB,
and taking the write lock causes deadlocks during VE saves: the POST
to action=visualeditoredit conflicts with the POST to the recursive
request to the Parsoid REST API.  We can't use a GET request for
these requests without hitting query-length limits.

This patch allows REST API calls to set the `X-MediaWiki-Read` header
to opt-out of the SQLite obligatory lock.  This avoids the deadlock,
while still allowing the API call to use a POST and avoid query length
limits.

Bug: T259685
Change-Id: If37dc890a24a45c3a914e310b5b5bf625965e9e6
(cherry picked from commit b75ac3953e750fd6b1b29868a77dbebd7969fbdc)