cache: Add purges from HtmlCacheUpdater for NS_MEDIAWIKI js/css pages
Follows-up 6e2be46e828 (Idaa8552cf37), which allowed canonical
action=raw URLs to be long-cached by the CDN and have them be
purged on-edit alongside the view and history URLs.

In reviewing that, I mistook isJsSubpage() as checking both
NS_USER and NS_MEDIAWIKI js pages. This method name was renamed
to isUserJsConfigPage() which made the omission more obvious.

Bug: T261371
Change-Id: Ic22b5f39ce4e1a9f074833f9e4e626e3999465dd