mediawiki.language,preferences: Use RegExp.test instead of String.match
It's very common to use match() as if it would return a boolean.
But it doesn't. It returns, well, a match (as an array), or null
if no match was found.

test() is more narrow and simply returns a boolean.

Let's use the method that is meant to be used in a boolean
context. It reflects better what the code is meant to do, and is
potentially faster.

Change-Id: Ic57d791fb50a9670f66a45257c671fd72869490f