resourceloader: Fix incorrect order of feature stylesheets
Follow up to I755e5e6784481b419e35, which used array_unshift
to prepend the 'feature' stylesheets. This works as expected when
there is only one 'feature' enabled.

When there are multiple, use of unshift will effectivel reverse
the order as it unshifts then one at a time.

To mitigate this, collect them normally in the correct order,
and then prepend them all at once with array_merge.

Bug: T262507
Change-Id: Ibe2c9f8d024f6be06588a59df10a37681b60d6bc