Move notification area to end of body
This is a second take at moving the mw.notify area to be outside
the main $content element. It moves .mw-notification-area to be
at the end of the body, and adjusts its top offset to match that
of $content or (if a skin has provided its own area) the area.

This introduces a breaking change for skins that don't set their
own area and which assume that the starting offset of the area
matches that of $content. We've talked about this and it feels
more correct to do it this way and then have skins correct their
own display where required (Timeless for example already adds
an offset; we'll make patches for other skins as required).

Bug: T260413
Bug: T260262
Bug: T260432
Change-Id: I12db7c1745d17851f6421123a145de4cb916ad51