Fix GuzzleHttpRequest request headers
The headers passed to Guzzle via the Client constructor are termed
"conditional", meaning that headers in the Request override them. If you
set a "body" option containing a MultipartStream, it sets a default
Content-Type request header, overriding the one from the Client
constructor. So to send a multipart stream with a custom Content-Type,
it is necessary to pass headers to Guzzle in the Request object.

Change-Id: Ie7bfc548ce4211fbf3ec67e3913bbdd3482977c3