Conform to selector-class-pattern for watchlist expiry class
This is a follow up to I9aba8cccc2ef75c59b78008723d0ae76d853c5f2
to add the 'mw-' prefix to the CSS class added there.

The whole of special.less has the stylelint selector-class-pattern
rule being excluded, so this wasn't picked up when it was first
added. For this reason, the disabling is moved to next-line form
and copied to all the remaining incorrect selectors (to avoid
this problem in the future).

Change-Id: Idfaec3ff3aa0e3ecc5217aa4a557a67378991757