Fix failure of rebuildLocalisationCache.php due to RL hook
rebuildLocalisationCache.php can fail when clearing the MessageBlobStore
due to a hook while constructing a ResourceLoader. This problem was also
encountered in T231866, but the fix was only applied to
ServiceWiring.php and not this duplicated code in
rebuildLocalisationCache.php.

So, fix rebuildLocalisationCache.php to similarly not construct a
ResourceLoader object.

Bug: T262900
Change-Id: I4e28c0f4e06efd219226154ab955395e00be504f