[tests] Remove SpecialUnblock::processUnblock calls in SpecialUnblockTest
This commit deliberately does not add
any tests for the old, processUnblock
method, for the following reasons:

* The method is not actually used anywhere
  in Gerrit (according to Code search), so there
  is no proof of this deprecated functionality
  to be used in any extension. While it may be used
  in unpublished code, that should be rare enough,
  covered by the service's tests. The processUnblock
  method is also extremely trivial, and the probability
  of making a mistake directly in it (so it wouldn't
  be noticed by service's tests) is little to none.
* Even this patch technically decreased the
  test coverage, it actually increases it a lot.
  Prior to this commit, if someone makes a bug
  within SpecialUnblock (ie. outside of UnblockUser
  command object), it would go unnoticed by our tests.
  That is because our current coverage of
  the special page does not execute the special page
  itself (comparably to what a real request would do),
  it calls processUnblock method directly. That method,
  however, is not called by the special page at all.

I believe that accepting a very slight lowering
of the test coverage (basically only for an academical
case only) is only a minor and temporary disadvantage.

This method will be soon (in next release) removed
for good, so this doesn't really contribute to
the long-term technical debt.

A proper long term solution for similar issues would
be to write proper unit tests for this special page.

Bug: T250020
Change-Id: I8223420989edab2a20725425531d2d5462a4f32a