Installer: consistently reset Language objects
Language objects contain references to service objects, so they need to
be be discarded when the global service container is reset by calling
MediaWikiServices::resetGlobalInstace(). The Installer class was doing
this correctly in the constructor, however, there where other calls to
resetGlobalInstace() during the installation process that did not
properly reset all language objects, which resulted in incorrect instances
of LocalisationCache being accessed.

This patch introduces the resetMediaWikiServices() method into the
Installer class which can be used to safely reset the service container
during installation.

Bug: T261081
Change-Id: I0ccab2b7fcac6942016430b1df5a1dbeb192f2f6