Use the same character set for link prefix as for suffix in Arabic 'ar'
The default is a-z plus every non-ASCII character, but this
is too broad.

Instead use the same character set as is used for link trails,
specifically Latin & Arabic letters.

Bonus:
Add combining diacritics to both sets as when these are appended
to letters, the resulting glyph is still considered a letter.

Bug: T263266
Change-Id: I358673f79989491799d3d68da17e73b806b167e0