Introduce BlockUtils service
Right now, this service contains only target validation,
but it will soon be expanded with additional methods, cf. T263249
or T250737 in general.

I deliberately do not update usages of this method in core in this
patch, I plan to do that in a followup patch to make reviewing easier.

Bug: T263189
Change-Id: I69df8ac927fe38d64c3d8c1a8e2ed63fd12202b7