mediawiki.page.watch.ajax: resolve notif promise, then update watch link
With $wgWatchlistExpiry set, OOUI is loaded when showing the
notification popup. This commit moves the updating of the watch link to
happen after the mw.notify promise is resolved, to avoid race conditions
with extensions that assume notifications are shown before updating the
link.

For T263135 specifically, this effectively means we remove all support
for watchlist expiry from StructuredDiscussions, but we had not indended
adding support anyway as the extension is deprecated. Morever, it is
safer to maintain the same execution flow in the event there are other
extensions that intercept mw.notify in a similar manner.

Bug: T263135
Change-Id: I5dd62c7b0ba386c3dcb826a1863adb7b97a455bd