Provide basic anon personal menu placeholder.
With this minimal change, the placeholder is integrated into the
personal tools. It's basic by default, so skins can use it as is,
customize it further if need be or even unset it altogether.
For instance, Vector uses different icon from what from MonoBook
uses, but both need to duplicate the basic placeholder.

Timeless needs to unset it. Modern can use it as is.

The benefit is that this will allow us to remove the hacky string
concatenation for it in Vector and also remove the duplication for
the same thing in MonoBook. 

More importantly, it will also allow us to remove a hack for the
UniversalLanguageSelector extension inside at least three skins
which was necessitated because the string concatenation, besides
being isolated from the menu building abstraction, is also
interfering with some permutations about the position of the items.

Note anonymous pages are subject to caching you please apply
?action=purge when testing.

Bug: T263382
Depends-On: I5691529ab8c59f4053cff38ea6f7dd01c326c074
Depends-On: I008511f5fc326bca10f2c2bf4e03b2df3561b56e
Depends-On: I0ba502a1e1368c1080caee2ce2bb45c27fdf77c3
Change-Id: Ice379cd8a68530b586a8a8c288ac65bb978f775c