Ignore block restrictions if block is not specified as partial
In SpecialBlock, ignore restrictions if the 'sitewide' radio was
selected. In ApiBlock, ignore restrictions if the 'partial' parameter
was falsey.

In both cases, ignoring restrictions means not passing them to
BlockUser.

This restores the behaviour prior to a656d03597d2, that a block is
sitewide unless it is marked as partial, even if restrictions options
are set.

Bug: T189073
Change-Id: I4d2e08b744b9d59909ce24cc5f167b676478970d