Search files in user language
The uselang will not matter to the response: there are
no localizable data, other than imageinfo, which will
be set separately (via iiextmetadatalanguage, in
mw.widgets.MediaResourceProvider)

But it might matter for how the search term is interpreted.
While the language of a search term has not traditionally
mattered, it now does when structured data is also being
considered. E.g. a German user searching for "gift" is
probably expecting files depicting Q40867 (poison), while
English users are probably looking for Q184303.

So while this won't change the language of whatever data
is returned, it does inform the search backend about the
intent of the search.

Bug: T259896
Change-Id: I64283967551b9463b5d85be3785f82f5863a67e7