ResourceLoaderSkinModule: Move edit link into interface feature
While working on the Example skin we noticed that the edit section
links and hiding of the print footer are important styles.

Rather than making the Example skin inherit the legacy feature, these
essential rules are added to the existing interface styles.

For backwards compatibility, legacy continues to ship the interface
edit links.

For skins loading both interface and legacy, the styles will be loaded
twice but we will lean on gzip to avoid an increase in CSS bytes

Bug: T89981
Change-Id: I7d35b44fc8c135320dfd1d99b16ba47ebaac876b