LivePreview: Show "Templates used in this preview" when creating a new page or editing a section
mediawiki.action.edit.preview.js now reconstructs "Templates used in
this preview" beneath the edit box if it doesn't exist, so that the list
of transclusions in the preview returned by the Parse API will now be
available to the user even when creating a new page or editing a
section. It also now correctly replaces the label "Templates used on
this page" with "Templates used in this preview", or "Template" if there
is only one, thus more closely matching the default PHP behavior.

Bug: T102286
Change-Id: I9976067d9c3a42d2a93d56da677354a8f352eb2f