Print footer should be hidden by default in all skins using `interface`
This rule is currently present in the legacy.less stylesheet but
skins should not be required to set the legacy feature.

All skins output a .printfooter element and it should always be hidden
for media screen.

Given its a single rule, we will live with the duplication noting
that the eventual goal is to remove the need for legacy altogether.

Note that the rule in src/mediawiki.skinning/commonPrint.css which sets
padding does not seem as essential and could be simplified
so is not moved to the interface feature (for now).

Bug: T89981
Change-Id: Ie14cf0e5555afb0fbc9d43014bce9e6e46287824