ApiPageSet: Avoid infinite loop when merging redirects
Redirects can form a loop but it was unaccounted for in this process.
Track seen pages to detect when a loop has formed. In the case of a
redirect loop the ApiPageSet will include the source page and ignore
the found redirects.

Bug: T264534
Change-Id: Ia5f4eeb9a4d90f02aceec3ed83bd4fc4a8a23ca4
(cherry picked from commit 861293a36cdbb2398f726c8fb5fe3df01717537b)