Add client-side performance metrics for legacy search
As part of comparing Vue search with legacy search, we need to track how
long it takes a keypress to load and render search results for legacy
search (Vue search will perform the same metrics).

This commit adds the following metrics which will only be used in our
synthetic tests. We are not collecting RUM metrics at this time.

Marks:

* mwVectorLegacySearchLoadStart: Marks start of loading the search
module.

* mwVectorLegacySearchLoadEnd: Marks the end of loading the search
module.

Measures:

* mwVectorLegacySearchLoadStartToLoadEnd: Measures the time it takes to
load the search module.

* mwVectorLegacySearchLoadStartToFirstRender: Measures the time it takes
from the start of loading the search module to the first render of results.

* mwVectorLegacySearchQueryToRender: Measures the time it takes from
the start of the fetch to the render of search results.

Additional changes:

* Fixed a bug in loadSearchModule that added an unneeded `focusin` event
listener when the input has already received focus.

Bug: T251544
Change-Id: I0fa6b8904bd43c87a68e9161f00d686a0e588966