objectcache: fix WANCache getWithSetCallback() when TTL < "lowTTL"
Previously, the refresh probability ramp-up logic assumed that
keys were always assigned logical TTLs >= "lowTTL". Thus, if a
key was saved with a TTL of 3 seconds and "lowTTL" was 10 seconds,
then the probability of a refresh at t=0 seconds would incorrectly
equal the value it should have at t=7.

Also:
* Use class constants for key metadata array keys. This makes it
  easy to track the usage of options and is less prone to typos.
* Improve some documentation slightly.

Bug: T264787
Change-Id: I7267e8639617fb8dc0850530465ff9d1b899b592