Introduce EditConstraint system, migrate unicode check to constraint
This is the start of "Phase 1" of the migration described in
T157658#6191234. For now, only a single constraint, the
requirement that EditPage::$unicodeCheck match the expected
value, is moved to the new system, in the UnicodeConstraint class.

Both the EditConstraintRunner and the new UnicodeConstraint
have 100% test coverage with pure unit tests.

Bug: T157658
Change-Id: I4fc806ec0fd631e265948a59244295eebec691fa