MovePage: Handle target page deletion failure gracefully.
Throwing exception is not appropriate here and will prevent proper usage
of ArticleDeleteHook, which is ostensibly meant to allow extensions to
cleanly abort the deletion if they need to.

When the target page deletion fails, instead of blowing everything up
with exception, just cancel the open atomic section and return the
user-friendly deletion error back to the user.

Bug: T265779
Change-Id: I66b4458a103f1274715e22a784344c55a62a59ac