Drop all "DROP SEQUENCE" in Postgres
tables.sql never was meant as an idempotent script or something that can
be ran on an existing database or schema. This serves no purposes and
can cause issues if tables-generated.sql gets to run first (by dropping
sequences that just got created), currently this doesn't happen and
installer, first installs manual tables but it can change in the future.

Change-Id: I299a0ae1e49f415910892981c07cfd582d198598