Ensure FormatMetadata::makeFormattedData always escapes EXIF values
This is a step in addressing T266707, by ensuring that every path
through FormatMetadata::makeFormattedData() passes the value through
either a specific formatting method, or else FormatMetadata::literal(),
which should properly wikitext-escape the value.

However, we don't actually want to fully-escape literals, because we
want URL auto-linking to still work. So in this patch, ::literal() is
still using htmlspecialchars() (like the original code did, at least
for the times it remembered to escape strings). That lets some
wikitext meta-characters through. Tightening up
FormatMetadata::literal() is left to a future patch, because it might
break some media pages if editors are deliberately (ab)using wikitext
markup here.

Bug: T266707
Change-Id: I289cf7687565612f8988a5c8a8211386e7904a44