Deprecate FormatMetadata::flattenArrayContentLang()
There appear to be no callers to flattenArrayContentLang() in code
search.  It also takes a $noHtml parameter which is labeled as a
backward-compat hack and we'd like to get rid of, in the case where
some caller for flattenArrayContentLang() does turn up.

Also mark FormatMetadata::flattenArrayReal() as @internal.  It's a
poorly named method and no one outside the media-handling code should
need to use it.

Change-Id: Ie3296d115816280740b5c182e7d0e2df20a4f909