Remove unused properties in SpecialComparePages
I couldn't find where they're being used currently in the class.
It seems $opts and $skin were supposed to be removed in this 2011 commit
that removed their usages: r79752 (128f33aa5edc)

For $showNavigation, it's unclear whether it has ever been useful.

No class is extending this class, so it's not possible for them to
be used anywhere else.

Change-Id: I792c26c270739a81ee15abc60c601b5c7def59d1