Extend FormatMetadata to handle non-numeric EXIF tags from PDF XMP data
The \Wikimedia\XMPReader\Info class defines the set of EXIF tags
parsed from PDF XMP data.  It includes a number of EXIF tags not
previously handled by FormatMetadata::makeFormattedData().  These are
EXIF tags, not PDF-specific tags, so they ought to be handled here in
core not in the PDFHandler extension.

Numeric tags are handled via the default fallthrough, but add cases
for structured and string data to ensure we don't try to coerce these
values to numbers.

Bug: T266677
Change-Id: Ifc0bb1eda4c4e390014dae64500bdb01b5c6a2e5