[ApiUserrightsTest] Start with empty $wgAddGroups/$wgRemoveGroups
In some tests mergeMwGlobalArrayValue should merge the empty array,
but when have settings in LocalSettings.php the result is not the empty
array as needed for the tests.
Just start with empty array to work on top of that.

1) ApiUserrightsTest::testAddAndRemoveGroups with data set "Add with
only remove permission" (array(array(), array('sysop')),
array(array('sysop'), array()), array('bot'))
Failed asserting that two arrays are identical.
--- Expected
+++ Actual
@@ @@
 Array &0 (
     0 => 'bot'
+    1 => 'sysop'
 )
with $wgAddGroups['bureaucrat'] = true; in LocalSettings.php

Change-Id: I092a3353aa65d53c84b765c04ec213ed8294b65d