SECURITY: Set a dummy title for Action buttons on Special:Log
If no title is set, the fallback is to use the main page,
as defined by MediaWiki:Mainpage
On wikis where the main page can be translated, MediaWiki:Mainpage
is sometimes set to use Special:MyLanguage to redirect to the
appropriate version. This is interpreted as a special page, and
ACtion::getActionName forces the action to be 'view' if the title
cannot be used as a WikiPage, which includes all pages in NS_SPECIAL.
Set a dummy title to avoid this. The title provided is unused
by the SpecialPageAction class and does not matter.

Bug: T205908
Change-Id: I0f0d2e592df96c9bad9001baede900f716433cfa