MysqlUpdater: Add utility functions to alter default constraint
There's occassionaly a need to alter default constraints, currently
a separate patch file is required to do so everytime which is quite
onerous and can be simplified.

Borrow a pair of set/drop functions from PostgresUpdater to make the
process more efficient.

Also convert usages of patch files to use the methods.

Change-Id: I846b94d0a721295f5bcdfe21f658116704e2d350