Remove unnecessary $wgHooks reference
The hook is not called because the script calls setTemplateCallback(),
avoiding statelessFetchTemplate() which calls the hook.

The hook function is wrong anyway and wouldn't have worked if it was
called.

Removing because it makes Phan fail on an unrelated change.

Change-Id: Ia6a5b265dca5e0debc6f66dd5287105f44bbde88
(cherry picked from commit 3aa7afb2657ebfc6ceb20165f382dd239af76c8e)