Use "迭代" instead of "疊代" for zh-TW
"Iterate" should be "迭代" according to the National Academy for
Education Research of Taiwan [1]. It's based on the meaning of "repeatedly;
again and again" of the word "迭" [2]. Although "迭" sometimes also share
the same meaning as "疊" but only when it means "to fold; to fold up; to fold
over in layers; to pile up; to be piled up; etc." [2].

This change only changes $zh2TW because I cannot find authoritative HK
documents for the usage of "迭代". However, I still suggest that we should
also change $zh2HK in the future.

[1] http://terms.naer.edu.tw/detail/10575858/
[2] https://en.wiktionary.org/wiki/%E8%BF%AD