Drop table of contents test assertions around visibility
The collapsibility/visibility of the table of contents is a skin decision. A skin for
example may choose to not make the table of contents collapsible - in
fact this is planned for Vector in a future version:
https://www.mediawiki.org/wiki/Reading/Web/Desktop_Improvements/Features#Table_of_contents

The test no longer makes sense in core so the check for visibility
is removed here. Vector will be a better location for the test if we want
to retain this for the legacy skins.

Bug:  T252774
Change-Id: I876ce99363d6c594516c983bbfac9db7cbe770be