language: Clean up $separatorTransformTable in km/la/my
LanguageKm and LanguageMy defined overrides for ::formatNum()
which did nothing -- but what they really wanted to do was just to
suppress separators.  Do that in the 'modern' way by invoking
::formatNumNoSeparators().

MessagesLa and MessagesKm defined $separatorTransformTable for one but
not both of the keys `.` and `,`.  Add defaults to
Language::formatNum() to handle this case without burping out a PHP
notice.  Use belt and suspenders by also defining an identity mapping
for '.' in MessagesLa::$separatorTransformTable.

Bug: T267091
Change-Id: I0169606ca1e211d241fa71f23ee0a16edc64b7ae